refactor: unify route client delegate filename to page-client
detail routes used client.tsx while list routes used page-client.tsx for
the same role (the "use client" component delegated to by a server
page.tsx). rename the three detail client.tsx files to page-client.tsx,
update sibling imports, and document the convention in CLAUDE.md.
