+++ src/app/(i18n)/about/[id]/page-client.tsx
