+++ src/app/(i18n)/blog/[idx]/page-client.tsx
