+++ src/app/(i18n)/recruit/[idx]/page-client.tsx
